remove some warnings

+++ lib/adore_controllers/include/controllers/iLQR.hpp
@@ -34,7 +34,7 @@ class iLQR
 {
 private:
 
-  int    horizon_steps         = 30;  // Number of time steps
+  size_t horizon_steps         = 30;  // Number of time steps
   double dt                    = 0.1; // Time step duration
   double heading_weight        = 10.0;
   double lateral_weight        = 50.0;
@@ -45,7 +45,7 @@ private:
   double jerk_weight           = 0.1;
   double steer_rate_weight     = 0.1;
   double convergence_threshold = 1e-6;
-  int    max_iterations        = 10000;
+  size_t max_iterations        = 10000;
 
   // Initialize variables for the backward pass
   std::vector<Eigen::MatrixXd> A_list;
